test(mcp-server): wiremock integration test for tool dispatch and auth passthrough
Refactors build_client / build_router out of main.rs into a lib target
so the integration test can call them directly without duplicating the
~85-tool registration list.
