+++ crates/superposition_mcp_server/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod auth;
+pub mod build;
+pub mod config;
+pub mod dispatch;
